feat(InputValidation): Refactor to separate input validation into its own class
